added functionalities to all the icons
