--- frontend/src/components/FileUploader/Uploader.css
@@ -1,39 +1,0 @@
-.buttons{
-    width: 200px;
-    height: 50%;
-    position: relative;
-    display: flex;
-    /* overflow: hidden; */
-    align-items: center;
-    align-self : center;
-    margin : 0;
-    align-content: center;
-    left: 50%; 
-    transform: translateX(-50%);
-    padding-top: 3px;
-    padding-bottom: 3px;
-    padding-left: 50%;
-    padding-right: 50%;
-    border-radius : 10px;
-    justify-content: center;
-    width: 45%;
-    size-adjust: 10px;
-    align-items: left;
-    z-index: 10%;
-}
-
-.icon{
-    position: relative;
-    flex-wrap: wrap;
-    flex-direction: row;
-    border-radius: 10%;
-    margin-right : 10px;
-    width: 6%;
-    height: 5%;
-    background: transparent;
-    /* margin-left: 30px; */
-}
-
-.images{
-    height : 18px;
-}
