+++ frontend/src/components/PostBox/PostBox.js
@@ -1,55 +1,118 @@
 import React from 'react';
-import { Container, Button, Dropdown, DropdownButton} from "react-bootstrap";
+import { Container, Button} from "react-bootstrap";
 import "./PostBox.css";
-import PhotoUploader from '../FileUploader/PhotoUploader';
-import PdfUploader from '../FileUploader/PdfUploader';
-import LinkUploader from '../FileUploader/LinkUploader';
-import { useState, useEffect } from 'react';
-import sicon from "../../images/send-outline.svg";
-import FileBase from 'react-file-base64';
-import axios from 'axios';
 
+import { useState} from 'react';
+import sicon from "../../images/send-outline.svg";
+import licon from "../../images/link-outline.svg";
+import ficon from "../../images/document-attach-outline.svg";
+import styled from 'styled-components';
+import picon from "../../images/camera-outline.svg";
+
+const Button1 = styled.button`
+  position: relative;
+  flex-wrap: wrap;
+  flex-direction: row;
+  border-radius: 10%;
+  margin-right : 10px;
+  width: 6%;
+  height: 5%;
+  background: transparent;
+`;
 
 function PostBox(props){
 
-        const [postData, setPostData] = useState({ content : '', tag : '', attachments : ''});
+        const [postData, setPostData] = useState({ createdBy : '1', content : '', tag : '', attachments : ''});  
 
-        const onPostClick = () => {
+        const getBase64 = async (e) => {
+            const file = e.target.files[0];
+            const base64 = await convertBase64(file);
+            return base64;
+          }
+        
+        const convertBase64  = (file) => {
+            return new Promise((resolve, reject) =>{
+              const fileReader = new FileReader();
+              fileReader.readAsDataURL(file);
+              fileReader.onload = () => {
+                resolve(fileReader.result);
+              };
+              fileReader.onerror = (error) => {
+                console.log(error);
+              };
+            });
+          }
 
-        }   
+        const hiddenFileInput = React.useRef(null);
 
-        const dropSelect = (e) => {
-            var val = e.target.value;
-            if(val == null){
-                console.log("error! Tag Mandatory!");
-            }
-            else{
-                return val;
-            }
+        const handleClick = event => {
+            hiddenFileInput.current.click();
+        };
+
+        function handleSubmit(){
+            fetch('localhost:8080/post', {
+                method : 'post', 
+                headers:{'Content-type':'application/json'},
+                body:JSON.stringify({postData})
+            }).then(response => response.json()).then(data => {
+                console.log(data);
+                setPostData({ createdBy : '1', content : '', tag : '', attachments : ''});
+            }).catch(error => console.log('error', error))
         }
 
         return(
             <Container className = "PostBox">
                 <div >
-                    <input className = "commentBox" placeholder = "What's on your mind?"  onChange = { (e) => setPostData({...postData, content : e.target.value})}/>
+                    <input className = "commentBox" placeholder = "What's on your mind?"  onClick = { (e) => setPostData({...postData, content : e.target.value})}/>
                 </div>    
                 <div className = "buttons">
                     <div className ="dropdown">
-                            <select className = "dropbtn">
+                            <select className = "dropbtn" onClick = {(e) => setPostData({...postData, tag : e.target.value})}>
+                                <option value = "1">Select Tag</option> 
                                 <option value = "Job-Recruitment">Job-Recruitment</option>
                                 <option value = "Knowledge Sharing">Knowledge Sharing</option>
                                 <option value = "Inspiration">Inspiration</option>
-                                <option value = "Others">Others</option>
-                                <option value = "Select Tag" selected >Select Tag</option>
+                                <option value = "Others">Others</option>    
                             </select>
                     </div>
-                    <PhotoUploader type="file" multiple={false} onDone={ props => console.log(props.base64)} />
-                    <PdfUploader type="file" multiple={false} onDone={ props => console.log(props.base64)} />
-                    <LinkUploader type="file" multiple={false} onDone={ props => console.log(props.base64)} />
 
-                    <Button className = "icon">
+                    <>
+                        <Button1 onClick={handleClick}>
+                            <img src = {picon} className = "images" alt = ""/>
+                        </Button1>
+                        <input
+                            type="file"
+                            ref = {hiddenFileInput}
+                            style={{display: 'none'}}
+                            onChange = { (e) => setPostData({...postData, attachments : getBase64(e)})}/>
+                    </>
+
+
+                    <>
+                        <Button1 onClick={handleClick}>
+                            <img src = {ficon} className = "images" alt = ""/>
+                        </Button1>
+                        <input
+                            type="file"
+                            ref={hiddenFileInput}
+                            onChange = { (e) => setPostData({...postData, attachments : getBase64(e)})}
+                            style={{display: 'none'}}
+                        />
+                    </>
+                    <>
+                        <Button1 onClick={handleClick}>
+                            <img src = {licon} className = "images" alt = ""/>
+                        </Button1>
+                        <input
+                            type="file"
+                            ref={hiddenFileInput}
+                            onChange = { (e) => setPostData({...postData, attachments : getBase64(e)})}
+                            style={{display: 'none'}}
+                        />
+                    </>
+
+                    <Button className = "icon" type = "submit" onClick={handleSubmit}>
                         <img src = {sicon} className = "images" alt = "" />
-                        {/* <span>post</span> */}
                     </Button>  
                 </div>         
             </Container>
